refactored tui section. Created buildview
